+++ src/components/Buttons/index.ts
@@ -1,2 +1,2 @@
 export * from './Button/Button';
-export * from './CopyBtn/CopyBtn';
+export * from './CopyButton/CopyButton';
